backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]==0.115.12",
    "uvicorn[standard]==0.24.0",
    "sqlalchemy[asyncio]==2.0.37",
    "alembic>=1.14.1",
    "pydantic-settings==2.7.1",
    "asyncpg>=0.30.0",
]
